+++ clubcorner/src/pages/home/home.html
@@ -24,18 +24,9 @@
 
       
 
-
-
-  <ion-fab left bottom>
-      <button ion-fab color="dark"><ion-icon name="arrow-dropup"></ion-icon></button>
-      <ion-fab-list side="top">
-        <button ion-button round small color="dark" style="float:left;" (click)="openPlayerInviteModal()">Spieler einladen</button>
-        <button ion-button round small color="dark" style="float:left;" (click)="openPlayerListModal()">Spielerliste</button>
-        <button ion-button round small color="dark" style="float:left;" (click)="openCreateTrainingModal()">Training planen</button>
-        <button ion-button round small color="dark" style="float:left;" (click)="openCreateGameModal()">Spiel anlegen</button>
-        <button ion-button round small color="dark" style="float:left;" (click)="showConfirm()">Mannschaft auflösen</button>
-      </ion-fab-list>
-    </ion-fab> 
+<button [navPush]="teamPage" [navParams]="{team:team}" ion-item>
+          TeamPage
+        </button> 
 
 
 </ion-content>
