+++ clubcorner/src/pages/home/home.ts
@@ -26,59 +26,13 @@ export class HomePage {
   allteams: Team[];
 
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, public modalCtrl: ModalController, public alertCtrl: AlertController, private _teamProv: Services) {
+  constructor(public navCtrl: NavController, public navParams: NavParams, private _teamProv: Services) {
     // If we navigated to this page, we will have an item available as a nav param
     this.selectedItem = navParams.get('item');
     this.getAllTeams();
   }
 
-  openPlayerInviteModal() {
-    let myModal = this.modalCtrl.create(PlayerInviteModalPage);
-    myModal.present();
-  };
-  
-  openPlayerListModal() {
-    let myModal = this.modalCtrl.create(PlayerListModalPage);
-    myModal.present();
-  };
 
-  openCreateGameModal() {
-    let myModal = this.modalCtrl.create(CreateGameModalPage);
-    myModal.present();
-  };
-
-  openCreateTeamModal() {
-    let myModal = this.modalCtrl.create(CreateTeamModalPage);
-    myModal.present();
-  };
-
-  openCreateTrainingModal() {
-    let myModal = this.modalCtrl.create(CreateTrainingModalPage);
-    myModal.present();
-  };
-
-
-  showConfirm() {
-    const confirm = this.alertCtrl.create({
-      title: 'Manschaft auflösen',
-      message: 'Sind sie sich sicher, dass sie die Mannschaft auflösen wollen?',
-      buttons: [
-        {
-          text: 'Ablehnen',
-          handler: () => {
-            console.log('Disagree clicked');
-          }
-        },
-        {
-          text: 'Zustimmen',
-          handler: () => {
-            console.log('Agree clicked');
-          }
-        }
-      ]
-    });
-    confirm.present();
-  }
 
 
 //-----------------------------------------------------------------
@@ -95,6 +49,8 @@ export class HomePage {
       error => console.log(error)
     )
   }
+
+
