+++ clubcorner/src/pages/modals/player-list-modal/player-list-modal.ts
@@ -21,4 +21,13 @@ export class PlayerListModalPage {
   closeModal() {
     this.viewCtrl.dismiss();
   }
+
+   /* createGameModalPage(){    
+    this._teamProv.createTermin(this.uebergabe).subscribe(
+      (data) => {
+        console.log(data);
+      },
+      error => console.log(error)
+    )
+  }*/
 }
