+++ clubcorner/src/pages/team/team.scss
@@ -43,5 +43,23 @@ page-team {
       }
 }
 
+page-team {
+    ion-fab-list {
+        -webkit-box-align: left !important;
+        -webkit-align-items: left !important;
+        -ms-flex-align: left !important;
+        align-items: flex-start !important;
+    }
+}
+body > ion-app > ng-component > ion-nav > page-team > ion-content > div.scroll-content > ion-list{
+    margin-left: 16px;
+    margin-right: 16px;
+}
+body > ion-app > ng-component > ion-nav > page-team > ion-content > div.scroll-content > ion-list > ion-item > div.item-inner > div > ion-input > input{
+    margin: 0px;
+  }
+  body > ion-app > ng-component > ion-nav > page-team > ion-content > div.scroll-content > ion-list > button:nth-child(1){
+      padding-left: 0px;
+  }
